pdfmake-test: minor changes
