+++ src/helpers/pdf.js
@@ -4,7 +4,6 @@ const fs = require('fs');
 const util = require('util');
 const handlebars = require('handlebars');
 const puppeteer = require('puppeteer');
-const axios = require('axios');
 const PdfPrinter = require('pdfmake');
 const getStream = require('get-stream');
 
@@ -60,18 +59,6 @@ exports.generatePdf = async (data, templateUrl, options = { format: 'A4', printB
   } catch (e) {
     if (browser) await browser.close();
     throw e;
-  }
-};
-
-exports.getBase64ImageFromURL = async (url) => {
-  try {
-    const image = await axios.get(url, { responseType: 'arraybuffer' });
-    const raw = Buffer.from(image.data).toString('base64');
-
-    return `data:${image.headers['content-type']};base64,${raw}`;
-  } catch (e) {
-    console.error(e);
-    return null;
   }
 };
